requirements.txt
websockets == 9.1
pyserial == 3.4
